Jmol.___JmolVersion="15.2.71" // P also 14.32.71
bug fix: isosurface slab designations "xy" "xz" "yz" "-xy", "-xz" "-yz"
etc. are not correct

new feature: isosurface slab designations "a=" "b=" "c=" "-a=" "-b="
"-c=" "ab" "bc" "ac" "-ab" "-bc" "-ac"
 -- like x, y, and z, but for unit cell axes
   for example:
   		load spacegroup 1 unitcell [3 3 3 60 60 60]
   		isosurface sphere 5.0 slab -a=1
 -- note that negative options and unit cell options are not documented

new feature: color property xxxx [quoted schemeFileNameOrURL]
 -- adds nine more standard color schemes, including five "universally
readable" scientific
 	palettes (batlow,cividis,kry,thermal,viridis) as discussed in
https://www.nature.com/articles/s41467-020-19160-7
    as well as four more (inferno,magma,plasma,turbo) that are from the
R-Project viridis package
    	https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html

    for example:

 	load *1crn
 	color property atomno "viridis"

 -- scheme palettes are located as *.lut.txt in the
src/org/jmol/util/colorschemes/ directory
 -- column format .lut.txt (three columns of r g b per line, either
0-255 or 0-1, or four columns where first column is the index 1-256)
 -- also reads files using https://... from any online source in binary
or ASCII formats:
 		binary [optional 32-bit header
ignored][rrrr...(256)][gggg...(256)][bbbb...(256)] 768 or 800 bytes
total
    	ASCIIt (either three columns of r g b per line, or four columns
index r g b)

    for example:

 	load *1crn
 	color property atomno
"https://github.com/fiji/fiji/blob/master/luts/cool.lut?raw=true"

 -- filenames also work in the color() function:

  print color("viridis").colors

  print color("https://github.com/Image-Py/imagepy/blob/master/imagepy/data/luts/Others/morgenstemning.lut?raw=true").colors

 -- for more color palette look-up table files, see
      https://github.com/fiji/fiji/tree/master/luts
      https://github.com/Image-Py/imagepy/tree/master/imagepy/data/luts
	  https://sites.imagej.net/NeuroCyto-LUTs/luts

+++ src/org/jmol/symmetry/UnitCell.java
@@ -797,10 +797,10 @@ class UnitCell extends SimpleUnitCell implements Cloneable {
             String s = atrans[j];
             int sfpt = s.indexOf("/");
             if (sfpt >= 0) {
-              ftrans[j] = PT.parseFloat(s.substring(0, sfpt))
-                  / PT.parseFloat(s.substring(sfpt + 1));
+              ftrans[j] = PT.parseDouble(s.substring(0, sfpt))
+                  / PT.parseDouble(s.substring(sfpt + 1));
             } else {
-              ftrans[j] = PT.parseFloat(s);
+              ftrans[j] = PT.parseDouble(s);
             }
           }
         P3d ptrans = P3d.new3(ftrans[0], ftrans[1], ftrans[2]);
